+++ src/BlacklistServiceProvider.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Milenmk\LaravelBlacklist;
 
 use Illuminate\Support\ServiceProvider;
+use Milenmk\LaravelBlacklist\Http\Middleware\BlacklistMiddleware;
+use Milenmk\LaravelBlacklist\Services\BlacklistService;
 
 class BlacklistServiceProvider extends ServiceProvider
 {
@@ -30,5 +32,10 @@ class BlacklistServiceProvider extends ServiceProvider
         $this->publishes([
             __DIR__ . '/../config/blacklist.php' => config_path('blacklist.php'),
         ], 'blacklist-config');
+
+        $this->app['router']->aliasMiddleware(
+            'blacklist',
+            BlacklistMiddleware::class
+        );
     }
 }