+++ tests/Feature/BlacklistServiceProviderTest.php
@@ -6,7 +6,7 @@ namespace Milenmk\LaravelBlacklist\Tests\Feature;
 
 use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Support\Facades\Config;
-use Milenmk\LaravelBlacklist\BlacklistService;
+use Milenmk\LaravelBlacklist\Services\BlacklistService;
 use Milenmk\LaravelBlacklist\Tests\BaseTest;
 use PHPUnit\Framework\Attributes\Test;